PickUp Code, and KalpiBus Enemy
I used a manager class for pickups in the end because I thought it was logicly correct to put them there and not inside the gamemanager class itself. I created the buss enemy, I used a polygon collider because the game isn't proccesing heavy and accuracy is needed. so you either miss or hit with extreme precision.
I tested the buss with diffrent kinds of colliders.
added mandats and popluarty floats to the game manager class.
that's about it.